Week2/requirements.txt
sentence-transformers
chromadb
openai
python-dotenv
nltk